Add PCL detector + Lidar detector
PCL: publish obstacle's centroid in cds_msgs/PclObstacle
Lidar: publish obstacles (line/circle) in cds_msgs/Obstacles

Remove CV3 requirements in data_collection

+++ data_collection/src/data_collection_node.cpp
@@ -75,7 +75,7 @@ void process()
     // cv::imshow("DepthRaw", depthImage);
 
     cv::Mat depthThresholded;
-    cv::inRange(depthImage, cv::Scalar{minDepth}, cv::Scalar{maxDepth}, depthThresholded);
+    cv::inRange(depthImage, cv::Scalar{double(minDepth)}, cv::Scalar{double(maxDepth)}, depthThresholded);
     // cv::imshow("DepthThresholded", depthThresholded);
 
     cv::Mat result;
